defines builder
